test: 优化测试用例并添加新功能的测试
- 优化现有测试用例，提高代码覆盖率
- 添加新功能的测试用例，包括日志记录、确认对话、重试机制等
- 增加对 web_search、visit_webpage 等默认工具的测试
- 改进对 llm 装饰器和提示生成相关函数的测试

+++ src/uglychain/llm.py
@@ -14,7 +14,7 @@ from .schema import Messages, P, T, ToolResponse
 from .structured import ResponseModel
 
 
-def retry(n: int, timeout: int, wait: int) -> Callable[[Callable[P, Any]], Callable[P, Any]]:
+def retry(n: int, timeout: float, wait: float) -> Callable[[Callable[P, Any]], Callable[P, Any]]:
     def decorator_retry(func: Callable[P, Any]) -> Callable[P, Any]:
         max_retries = n
         llm_timeout = timeout
